+++ MORE/docs/release-notes.md
@@ -7,6 +7,9 @@
 ### Abyssal Power event
 * The Abyssal Power card now puts the Abyssal Visions card in your draw pile instead of your hand.
 
+### Dracula's Deck Trial
+* Improved clarity of dialogue options.
+
 ## 1.4.0
 Released 2 March 2025.